v1.0.4a - Fix breaking changes

+++ functions.js
@@ -54,7 +54,7 @@ module.exports = {
       // Disable the buttons for the users
       await message.edit({ content: content, components: [row] });
       setTimeout(() => {
-        message.delete();
+        if(!message.deleted) message.delete();
       }, 10000);
       // Return the button
       return res;
